develop - additional clean up of gradle files and project wide package name refactor
