+++ app/src/main/java/com/bottlerocketstudios/mapsdemo/MainActivity.kt
@@ -1,5 +1,4 @@
-package com.bottlerocketstudios.android_maps_demo
-
+package com.bottlerocketstudios.mapsdemo
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -10,7 +9,7 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.bottlerocketstudios.android_maps_demo.ui.theme.AndroidMapsDemoTheme
+import com.bottlerocketstudios.mapsdemo.ui.theme.AndroidMapsDemoTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
