+++ app/src/main/java/com/bottlerocketstudios/mapsdemo/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.bottlerocketstudios.android_maps_demo.ui.theme
+package com.bottlerocketstudios.mapsdemo.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
